[renderblocking] Implement render-blocking for link headers
This patch implements render-blocking for preload and modulepreload link
headers. It's done by creating a PendingLinkPreload when trying to fetch
the resource, so that we can receive callbacks when the fetch finishes.

Since there's no element for a link header, the test case cannot listen
to the 'load' event of any EventTarget. Therefore, this patch also
modifies the test code to allow waiting for the PerformanceEntry for
the link header's url.

Bug: 1271296
Change-Id: Ic64d26341516a510079c01dbff0b0196c7ec9f24