Make color-scheme meta affect all normal schemes
The HTML specification changed to use the meta color-scheme as a page's
supported color schemes instead of presentation style for the root
element. This means the color-scheme meta no longer affects the
computed for an initial value root element, but still affects the used
color-scheme the same way as before.

There is one used value change, and that is when resetting the
color-scheme to "normal" down the tree, which now follows the meta
element instead of resolving to a "light" color-scheme.

We now align with Gecko and Safari.

See the whatwg issue[1] for a discussion.

[1] https://github.com/whatwg/html/issues/7213

Bug: 1260617
Change-Id: Ic72a2dead75257650f8d99cc5f30f13b7005ea41